Move galaxy.web.framework.webapp into galaxy-webapps because it use galaxy-app stuff.
