+++ lib/galaxy/webapps/galaxy/controllers/authnz.py
@@ -30,7 +30,7 @@ class OIDC(JSAppLauncher):
         GET /authnz/
             returns a list of third-party identities associated with the user.
 
-        :type  trans: galaxy.web.framework.webapp.GalaxyWebTransaction
+        :type  trans: galaxy.webapps.base.webapp.GalaxyWebTransaction
         :param trans: Galaxy web transaction.
 
         :param kwargs: empty dict
